feat(config): add AppConfig class and refactor existing configuration management
